feat: use abbreviated Thai month in shared bill summary date
Change the copy/share date from full month to short form
(e.g. "24 กรกฎาคม 2569" -> "24 ก.ค. 2569") via month: 'short'.
Bump the asset version to 20260725a so the change reaches cached clients.

+++ app.js
@@ -516,7 +516,7 @@ function buildBillSummaryText() {
     const now = new Date();
     let dateStr, timeStr;
     try {
-        dateStr = now.toLocaleDateString('th-TH', { day: 'numeric', month: 'long', year: 'numeric' });
+        dateStr = now.toLocaleDateString('th-TH', { day: 'numeric', month: 'short', year: 'numeric' });
         timeStr = now.toLocaleTimeString('th-TH', { hour: '2-digit', minute: '2-digit' });
     } catch (e) {
         dateStr = now.toLocaleDateString();
